requirements.txt
torch
numpy

git+https://github.com/openai/CLIP.git

opencv-python-headless
Pillow

# QD-DETR 의존성
scipy
